# Improve median_filter, rank_filter and percentile_filter