@grlee77 2020-08-18
This looks nice.

For the median over small kernels like 3x3, 5x5, 3x3x3, etc. you can get a large speedup using an optimized sorting network. I was playing with that in `cupyimg` awhile back. See for example the following case for an [optimal sorting network for a 5x5 kernel](https://github.com/mritools/cupyimg/blob/aed7f7cb5d6284bcbd5e7758b1a4f9a1f2741d5a/cupyimg/scipy/ndimage/_kernels/optimal_median_preambles.py#L129-L185).

In that file, I stored a dictionary with several custom sorting networks. If the kernel size [doesn't match one of those then it falls back to the general sorting algorithm](https://github.com/mritools/cupyimg/blob/aed7f7cb5d6284bcbd5e7758b1a4f9a1f2741d5a/cupyimg/scipy/ndimage/_kernels/filters.py#L447-L468).

The following benchmarks correspond to the code currently in cupyimg where there are special cases for 3x3, 5x5, 7x7 and 9x9 so those are faster (by a factor of 7.1, 21.5, 54.8 and 66.6 respectively to the general sort). 11x11, 13x13 and 15x15 are not faster as the general sort is still used for them.

If there is interest in such an approach, I am happy to adapt it for a PR here.

```
median_filter(3x3)  :    CPU:   32.566 us   +/- 1.408 (min:   31.111 / max:   35.021) us     GPU-0:  162.621 us   +/- 1.627 (min:  160.768 / max:  166.496) us
median_filter(5x5)  :    CPU:   33.686 us   +/- 3.036 (min:   30.689 / max:   41.201) us     GPU-0:  393.843 us   +/- 3.292 (min:  391.168 / max:  402.048) us
median_filter(7x7)  :    CPU:   33.263 us   +/- 3.050 (min:   30.890 / max:   40.326) us     GPU-0:  956.858 us   +/- 2.978 (min:  954.368 / max:  963.584) us
median_filter(9x9)  :    CPU:   34.397 us   +/- 2.358 (min:   32.163 / max:   40.467) us     GPU-0: 1929.754 us   +/- 4.096 (min: 1923.072 / max: 1936.384) us
median_filter(11x11):    CPU:  111.000 us   +/-21.506 (min:   80.191 / max:  152.505) us     GPU-0:161464.334 us   +/-411.832 (min:160932.861 / max:162075.363) us
median_filter(13x13):    CPU:  125.656 us   +/-19.744 (min:   81.292 / max:  156.323) us     GPU-0:311806.015 us   +/-660.781 (min:311108.612 / max:313030.670) us
median_filter(15x15):    CPU:  142.473 us   +/-17.857 (min:  102.500 / max:  161.238) us     GPU-0:550420.233 us   +/-1304.930 (min:548508.667 / max:551965.820) us
```