@anaruse 2020-08-19
Attaching the measured results of `percentile_filter` as this implementation is faster when percentile is closer to 0 or 100 😄 

* time-pf.py
```
import cupy
import cupyx
import cupyx.scipy
import cupyx.scipy.ndimage
w, h = 1920, 1080  # full HD
print('w, h = {}, {}'.format(w, h))
a = cupy.random.rand(w, h, dtype=cupy.float32)
for size in 3, 5, 7, 9, 11, 13, 15:
    for percentile in 12, 25, 50:
        name = 'percentile_filter(size={size}x{size},percentile={percentile})'.format(size=size, percentile=percentile)
        perf = cupyx.time.repeat(cupyx.scipy.ndimage.percentile_filter, (a, percentile, size), name=name, n_repeat=10, n_warmup=2)
        print(perf)
```

```
[ Current master branch ]
$ python time-pf.py 
w, h = 1920, 1080
percentile_filter(size=3x3,percentile=12):    CPU:  161.673 us   +/- 4.496 (min:  156.783 / max:  169.718) us     GPU-0:  532.896 us   +/- 4.690 (min:  527.680 / max:  541.440) us
percentile_filter(size=3x3,percentile=25):    CPU:  160.317 us   +/- 7.523 (min:  154.779 / max:  179.981) us     GPU-0:  531.453 us   +/- 7.696 (min:  525.440 / max:  551.168) us
percentile_filter(size=3x3,percentile=50):    CPU:  159.877 us   +/- 5.003 (min:  155.809 / max:  172.046) us     GPU-0:  531.418 us   +/- 4.749 (min:  527.072 / max:  543.008) us
percentile_filter(size=5x5,percentile=12):    CPU:  163.364 us   +/- 3.689 (min:  158.447 / max:  171.315) us     GPU-0: 1989.933 us   +/- 9.836 (min: 1977.344 / max: 2005.440) us
percentile_filter(size=5x5,percentile=25):    CPU:  161.485 us   +/- 3.508 (min:  156.395 / max:  168.946) us     GPU-0: 1986.746 us   +/-12.451 (min: 1970.880 / max: 2006.624) us
percentile_filter(size=5x5,percentile=50):    CPU:  161.382 us   +/- 3.277 (min:  157.122 / max:  167.714) us     GPU-0: 1987.357 us   +/-12.837 (min: 1962.272 / max: 2010.592) us
percentile_filter(size=7x7,percentile=12):    CPU:  165.959 us   +/- 4.234 (min:  162.024 / max:  176.757) us     GPU-0:15925.293 us   +/-28.931 (min:15855.072 / max:15957.984) us
percentile_filter(size=7x7,percentile=25):    CPU:  162.961 us   +/- 3.994 (min:  159.429 / max:  172.316) us     GPU-0:15763.606 us   +/-79.309 (min:15668.768 / max:15979.392) us
percentile_filter(size=7x7,percentile=50):    CPU:  162.102 us   +/- 2.955 (min:  159.487 / max:  167.849) us     GPU-0:15680.019 us   +/-52.756 (min:15609.856 / max:15769.440) us
percentile_filter(size=9x9,percentile=12):    CPU:  165.398 us   +/- 4.266 (min:  160.503 / max:  174.877) us     GPU-0:59959.562 us   +/-67.902 (min:59865.185 / max:60065.952) us
percentile_filter(size=9x9,percentile=25):    CPU:  163.487 us   +/- 3.152 (min:  159.403 / max:  168.211) us     GPU-0:59936.096 us   +/-47.709 (min:59852.577 / max:60007.648) us
percentile_filter(size=9x9,percentile=50):    CPU:  163.029 us   +/- 3.617 (min:  158.611 / max:  171.746) us     GPU-0:59958.957 us   +/-86.882 (min:59834.370 / max:60114.559) us
percentile_filter(size=11x11,percentile=12):    CPU:  169.756 us   +/- 5.402 (min:  163.103 / max:  180.319) us     GPU-0:119572.377 us   +/-99.822 (min:119428.383 / max:119715.073) us
percentile_filter(size=11x11,percentile=25):    CPU:  172.196 us   +/-14.830 (min:  161.926 / max:  207.929) us     GPU-0:119579.376 us   +/-114.314 (min:119343.811 / max:119802.078) us
percentile_filter(size=11x11,percentile=50):    CPU:  191.624 us   +/- 7.171 (min:  184.189 / max:  211.009) us     GPU-0:119555.837 us   +/-71.975 (min:119451.263 / max:119659.523) us
percentile_filter(size=13x13,percentile=12):    CPU:  195.229 us   +/-28.229 (min:  166.101 / max:  248.289) us     GPU-0:225439.673 us   +/-380.096 (min:224844.070 / max:225993.927) us
percentile_filter(size=13x13,percentile=25):    CPU:  170.105 us   +/- 5.310 (min:  164.404 / max:  180.763) us     GPU-0:225310.374 us   +/-317.875 (min:224838.531 / max:226029.160) us
percentile_filter(size=13x13,percentile=50):    CPU:  168.277 us   +/- 2.936 (min:  163.545 / max:  173.219) us     GPU-0:225211.743 us   +/-204.466 (min:224853.699 / max:225622.086) us
percentile_filter(size=15x15,percentile=12):    CPU:  173.079 us   +/- 4.722 (min:  168.212 / max:  181.682) us     GPU-0:385326.968 us   +/-240.899 (min:384921.478 / max:385689.056) us
percentile_filter(size=15x15,percentile=25):    CPU:  174.795 us   +/-17.137 (min:  164.494 / max:  224.121) us     GPU-0:385374.976 us   +/-328.882 (min:384974.731 / max:386147.095) us
percentile_filter(size=15x15,percentile=50):    CPU:  169.613 us   +/- 4.004 (min:  163.767 / max:  176.735) us     GPU-0:385426.718 us   +/-308.996 (min:384864.868 / max:386025.665) us

[ This PR ]
$ python time-pf.py 
w, h = 1920, 1080
percentile_filter(size=3x3,percentile=12):    CPU:  166.037 us   +/- 7.240 (min:  158.061 / max:  178.679) us     GPU-0:  331.990 us   +/- 7.362 (min:  323.712 / max:  344.608) us
percentile_filter(size=3x3,percentile=25):    CPU:  163.300 us   +/- 4.714 (min:  158.524 / max:  174.255) us     GPU-0:  352.083 us   +/- 4.600 (min:  347.808 / max:  362.720) us
percentile_filter(size=3x3,percentile=50):    CPU:  161.273 us   +/- 3.358 (min:  156.618 / max:  167.595) us     GPU-0:  358.518 us   +/- 3.537 (min:  353.760 / max:  365.152) us
percentile_filter(size=5x5,percentile=12):    CPU:  164.279 us   +/- 9.982 (min:  158.371 / max:  193.684) us     GPU-0:  719.773 us   +/-10.180 (min:  714.208 / max:  749.760) us
percentile_filter(size=5x5,percentile=25):    CPU:  162.209 us   +/- 3.390 (min:  157.961 / max:  168.222) us     GPU-0:  874.397 us   +/- 3.451 (min:  869.664 / max:  880.000) us
percentile_filter(size=5x5,percentile=50):    CPU:  162.206 us   +/- 2.868 (min:  159.862 / max:  168.489) us     GPU-0: 1025.318 us   +/- 6.300 (min: 1013.376 / max: 1035.776) us
percentile_filter(size=7x7,percentile=12):    CPU:  163.748 us   +/- 3.595 (min:  159.923 / max:  169.256) us     GPU-0: 1325.408 us   +/- 3.702 (min: 1321.280 / max: 1331.456) us
percentile_filter(size=7x7,percentile=25):    CPU:  163.199 us   +/- 3.217 (min:  159.425 / max:  170.117) us     GPU-0: 2367.517 us   +/- 7.874 (min: 2354.912 / max: 2380.992) us
percentile_filter(size=7x7,percentile=50):    CPU:  165.302 us   +/- 3.698 (min:  160.833 / max:  171.445) us     GPU-0: 6741.267 us   +/-28.516 (min: 6693.472 / max: 6787.680) us
percentile_filter(size=9x9,percentile=12):    CPU:  163.169 us   +/- 2.727 (min:  160.192 / max:  169.396) us     GPU-0: 3982.691 us   +/-12.214 (min: 3963.712 / max: 3995.840) us
percentile_filter(size=9x9,percentile=25):    CPU:  165.520 us   +/- 2.976 (min:  161.744 / max:  170.593) us     GPU-0:17845.779 us   +/-31.855 (min:17781.120 / max:17885.153) us
percentile_filter(size=9x9,percentile=50):    CPU:  167.101 us   +/- 3.693 (min:  161.795 / max:  173.740) us     GPU-0:32178.976 us   +/-27.620 (min:32134.880 / max:32219.776) us
percentile_filter(size=11x11,percentile=12):    CPU:  165.838 us   +/- 6.301 (min:  161.296 / max:  183.703) us     GPU-0:10374.998 us   +/-31.495 (min:10326.368 / max:10409.952) us
percentile_filter(size=11x11,percentile=25):    CPU:  168.525 us   +/- 4.621 (min:  161.926 / max:  175.393) us     GPU-0:49777.603 us   +/-33.984 (min:49723.167 / max:49826.241) us
percentile_filter(size=11x11,percentile=50):    CPU:  168.056 us   +/- 3.705 (min:  162.182 / max:  174.700) us     GPU-0:65107.482 us   +/-26.054 (min:65057.281 / max:65153.183) us
percentile_filter(size=13x13,percentile=12):    CPU:  166.082 us   +/- 3.465 (min:  160.277 / max:  172.353) us     GPU-0:28731.818 us   +/-151.612 (min:28488.319 / max:28988.512) us
percentile_filter(size=13x13,percentile=25):    CPU:  175.847 us   +/-11.798 (min:  167.925 / max:  206.563) us     GPU-0:92324.285 us   +/-28.760 (min:92254.883 / max:92362.083) us
percentile_filter(size=13x13,percentile=50):    CPU:  173.113 us   +/- 5.551 (min:  169.111 / max:  188.763) us     GPU-0:121240.797 us   +/-184.108 (min:121007.965 / max:121594.498) us
percentile_filter(size=15x15,percentile=12):    CPU:  167.802 us   +/- 1.963 (min:  164.351 / max:  170.595) us     GPU-0:30851.885 us   +/-158.239 (min:30698.433 / max:31226.561) us
percentile_filter(size=15x15,percentile=25):    CPU:  172.270 us   +/- 2.292 (min:  167.774 / max:  176.322) us     GPU-0:155071.115 us   +/-41.084 (min:154973.831 / max:155128.098) us
percentile_filter(size=15x15,percentile=50):    CPU:  172.583 us   +/- 2.307 (min:  170.458 / max:  178.069) us     GPU-0:168043.011 us   +/-162.920 (min:167679.459 / max:168319.839) us
```